+++ tcpClient/getEvents.js
@@ -12,9 +12,9 @@ module.exports = function(config) {
             assert(streamName, baseErr + 'Stream Name not provided');
 
             direction = direction || 'forward';
-            startPosition = startPosition == undefined && direction == 'backward' ? -1 : startPosition || 0;
+            startPosition = startPosition === undefined && direction === 'backward' ? -1 : startPosition || 0;
             length = length || 1000;
-            resolveLinkTos = resolveLinkTos == undefined ? true : resolveLinkTos;
+            resolveLinkTos = resolveLinkTos === undefined ? true : resolveLinkTos;
 
             if (length > 4096) {
                 console.warn('WARNING: Max event return limit exceeded. Using the max of 4096');
@@ -32,7 +32,7 @@ module.exports = function(config) {
                 return resolve(result.events);
             }
 
-            if (direction == 'forward')
+            if (direction === 'forward')
                 connection.readStreamEventsForward(streamName, startPosition, length, resolveLinkTos, false, null, config.credentials, handleResult);
             else
                 connection.readStreamEventsBackward(streamName, startPosition, length, resolveLinkTos, false, null, config.credentials, handleResult);