+++ tcpClient/subscribeToStreamFrom.js
@@ -1,9 +1,8 @@
-var debug = require('debug')('geteventstore:getAllStreamEvents'),
+var debug = require('debug')('geteventstore:subscribeToStreamFrom'),
     EventStoreClient = require('event-store-client'),
     createConnection = require('./createConnection'),
     Promise = require('bluebird'),
-    assert = require('assert'),
-    _ = require('lodash');
+    assert = require('assert');
 
 var baseErr = 'Subscribe to Stream From - ';
 
@@ -12,7 +11,7 @@ module.exports = function(config) {
         settings = settings || {};
         return new Promise(function(resolve, reject) {
             assert(streamName, baseErr + 'Stream Name not provided');
-            if (fromEventNumber == 0) fromEventNumber = undefined;
+            if (fromEventNumber === 0) fromEventNumber = undefined;
 
             var catchUpSettings = new EventStoreClient.CatchUpSubscription.Settings();
 
@@ -22,7 +21,8 @@ module.exports = function(config) {
             if (settings.debug) catchUpSettings.debug = settings.debug;
 
             var connection = createConnection(config, reject);
-            connection.subscribeToStreamFrom(streamName, fromEventNumber, config.credentials, onEventAppeared, onLiveProcessingStarted, onDropped, catchUpSettings);
+            var subscription = connection.subscribeToStreamFrom(streamName, fromEventNumber, config.credentials, onEventAppeared, onLiveProcessingStarted, onDropped, catchUpSettings);
+            debug('', 'Subscription:', subscription);
             resolve(connection);
         });
     };