+++ tests/_globalHooks.js
@@ -7,7 +7,7 @@ var eventstore;
 before(function() {
 	this.timeout(10 * 1000);
 
-	if (eventstore == undefined) {
+	if (eventstore === undefined) {
 		console.log('Starting in-mem ES...');
 
 		var intTcpPort = '--int-tcp-port=' + esConfig.options.intTcpPort;
@@ -29,7 +29,7 @@ before(function() {
 		return Promise.delay(4000);
 	}
 	return Promise.resolve();
-})
+});
 
 after(function(done) {
 	console.log('Killing in-mem ES...');