+++ tests/http.sendScavengeCommand.js
@@ -1,9 +1,7 @@
-var globalHooks = require('./_globalHooks');
+require('./_globalHooks');
 
 var httpConfig = require('./support/httpConfig');
-var assert = require('assert');
 var eventstore = require('../index.js');
-var uuid = require('uuid');
 
 describe('Http Client - Send Scavenge Command', function() {
     it('Should send scavenge command', function() {