+++ tests/http.writeEvents.js
@@ -1,8 +1,8 @@
-var globalHooks = require('./_globalHooks');
+require('./_globalHooks');
 
 var httpConfig = require('./support/httpConfig');
-var assert = require('assert');
 var eventstore = require('../index.js');
+var assert = require('assert');
 var uuid = require('uuid');
 
 describe('Http Client - Write Events', function() {
@@ -18,7 +18,7 @@ describe('Http Client - Write Events', function() {
             return client.getEvents(testStream).then(function(events) {
                 assert.equal(events[0].data.something, '456');
             });
-        })
+        });
     });
 
     it('Should not fail promise if no events provided', function() {
@@ -39,9 +39,9 @@ describe('Http Client - Write Events', function() {
 
         var testStream = 'TestStream-' + uuid.v4();
         return client.writeEvents(testStream, events).then(function() {
-            assert.fail('should not have succeeded')
+            assert.fail('should not have succeeded');
         }).catch(function(err) {
             assert(err, 'error expected');
-        })
+        });
     });
 });